network: Implemented ClientConnection
EGE::ClientConnection is an instance of Client on server side.
The client can send and receive abstract EGE::Packets. Every user of
network module must define its Packet (to define what data are sent
over the network and which way), Client (to define which Packets
it uses) and Server (to define which Clients it creates).

It will be some default EGE::EGEPacket, which will define engine's
internal protocol (with SceneObject etc. support)