egeNetwork: Add EGEGame and derive Client and Server from it
The EGEGame class stores Gameplay Object Registry and Scene that will
be used by Client or Server. The Scene must be set to work properly.
The GPOM is created by

We remember that the user must set Client as sub-loop in its
main loop (usually GUIGameLoop) to make Client fully working.
The Server can be used as standalone main loop, but you must
then derive your main loop class from it. In that case it's
better to make simple EventLoop and derive your main loop
from it (or derive GUIGameLoop and NOT create window on server-
-side).