scene: Add TilemapRenderer
The TilemapRenderer renders... tilemap that is passed to it
in constructor, using specified atlas (a texture with all tile types)
and atlas mapper (function that maps TileMap's TileType to atlas
texture coordinate).

It's now optimized for rendering only tiles visible on screen, but
does not implement render caching (pre-rendering tilemap in chunks
and later simply display it).