scene: Refactor SceneObject to use 3D coordinates :)
We don't have longer SceneObject2D and SceneObject as separate entities.
Now, everything is handled in SceneObject. 2D/3D separation is done by
Camera (which was renamed from CameraObject). This commit adds one type
of Camera: Plain2DCamera, which is equivalent of old CameraObject2D.

Advantages of this solution is that we can now easily do 2.5D or even 3D
scenes.

Note that egeNetwork and tilemaps are not working now. It will be adressed
in future commits (they need further refactoring anyway).

This commit fixes also some left build errors after changing VectorOperations
to be a normal "includable" file instead of Vector internal.