gui: Support creating multiple windows
This commit adds support for managing multiple windows using GUIGameLoop.
The new Window class was added to represent GUI window, which can carry
a GUIScreen and a renderer.

The Widget was refactored to take Window as an argument instead of
GUIGameLoop. Also, there are some new helper convenience methods
added to Widget, like getResourceManager().