gfx: Generalize support for shaders into GfxFilters
A lot of shaders needs additional information (e.g. screen size) to work.
This commit allows to set shaders and any other states to Renderer
before rendering a Renderable through GfxFilter::apply() method.

There are also implemented:
- ShaderFilter, which sets up a shader and passes a window size to it
- ShaderFilterWithRenderTexture, which renders scene on RenderTexture
and passes it as uniform to shader.

The pixelize shader seems not to work :(