Everywhere: Add new component system :)
This new system simplifies hierarchy and event handling, which are now
automatically passed to every child. It fully replaces EventLoop and
introduces uniform API for updates.

Note that not every object is fully supported; there are issues e.g. on
GUI/Scene boundary, but it's a start.

The next step will be actually handling that for e.g SceneObjects (for
compound SceneObjects)

This commit removes some features like:
* addWidget()
* onTimerTick() / onTimerFinish()
* CompoundWidget (which get merged into Widget)

It adds also:
* TickEvent (called every tick on every object).