event: Add basic loading/saving from file to KeybindManager :)
This commit adds support for loading/saving keybinds from file. The
default file name used is "keybinds.json".

For now, the data is static_casted right from SFML enums, this is not OK
because these may change in any moment and there is no check if the
loaded values are actually correct.